+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.21.2
 - fix: Fixes an issue if we have to deal with two types with the same name
 - feat: add support for @internal annotations
+- fix: Narrowing return types (generics) should not be considered a breaking change
 
 ## Version 0.21.1
 - fix: Fix analysis of pub cache packages